requirements.txt
pdfplumber==0.10.3
pandas==1.5.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
requests==2.31.0
beautifulsoup4==4.12.3
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.6.3 